WIP VTContext
Restructured backends to extend a top-level virtual object
	(more backends to update)

New VTProxy registration which just checkpoints status info
about that proxy, but also registers to the VTContext for special handling
of the actual data

  VeloC doesn't appear to have a way to checkpoint an object as
just a single node (e.g. each checkpoint of a proxy element is
a node's iteration's element data, not the iteration's element data),
so recovering to different node positions is currently not possible
  Working on that before I start the steps for handling recovery of
the VT data in the VTContext. I may just use update the file backend
to the new virtual hierarchy and use that for more configurability for now.

  Checkpoints of status info seem very large right now, some 66M for each node
with only one collection and one objgroup being checkpointed. I have not yet
investigated that

  Another todo, attaching listeners to collection element migration events,
to send the updated status info alongside. Not an immediate concern.

+++ examples/benchmark_multiviews.cpp
@@ -111,7 +111,7 @@ int main(int argc, char *argv[]) {
       }
 
       std::vector<Kokkos::View<double*>> MyViews(num_views);
-      for( int i = 0 ; i < num_views; ++i )
+      for( size_t i = 0 ; i < num_views; ++i )
       {
          Kokkos::resize( MyViews[i], M*nbLines);
          Kokkos::deep_copy(MyViews[i],1.0);
@@ -135,13 +135,13 @@ int main(int argc, char *argv[]) {
       }
 
       wtime = MPI_Wtime();
-      int i = 1 + KokkosResilience::latest_version(*ctx, "test_kokkos");
+      size_t i = 1 + KokkosResilience::latest_version(*ctx, "test_kokkos");
 
       while(i < nsteps) {
 
          KokkosResilience::checkpoint(*ctx, "test_kokkos", i, [=]() {   // Nic, tell me what should I put for []/
 
-         for( int j = 0; j < num_views; ++j )
+         for( size_t j = 0; j < num_views; ++j )
          {
 
             Kokkos::parallel_for( M*nbLines, KOKKOS_LAMBDA (const int& k )
